% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{bet}
\alias{bet}
\title{A bet object that is placed for a leg on a given camel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bet
}
\description{
A bet object that is placed for a leg on a given camel
}
\examples{
x <- bet$new("blue", 5)
}
\keyword{datasets}
