% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{board}
\alias{board}
\title{A board object on which the game is played}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
board
}
\description{
A board object on which the game is played
}
\examples{
y <- system$new(nPlayers = 2, players = c("Michael", "Alex"))
x <- board$new(y)
}
\keyword{datasets}
