% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{camel}
\alias{camel}
\title{Implements camel class based off of the board game pieces}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
camel
}
\description{
Implements camel class based off of the board game pieces
}
\examples{
x <- camel$new("blue", 1, 1)
}
\keyword{datasets}
