% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\name{cleanColors}
\alias{cleanColors}
\title{Correctly orders colors for graphing}
\usage{
cleanColors(subColors)
}
\arguments{
\item{subColors}{a subset of the full set of colors of camels}
}
\value{
colors in correct order for the game
}
\description{
Correctly orders colors for graphing
}
