% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{die}
\alias{die}
\title{A three sided die, assigned a color corresponding to a camel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
die
}
\description{
A three sided die, assigned a color corresponding to a camel
}
\examples{
x <- die$new("blue")
}
\keyword{datasets}
