% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiMethods.R
\name{maxPurseIncrease}
\alias{maxPurseIncrease}
\title{Maximize Purse Increase}
\usage{
maxPurseIncrease(relevantData)
}
\arguments{
\item{gameData}{}
}
\description{
Make a decision based on the current game board that maximizes the increase in the current
player's purse based on given historical data.
}
