% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{playCamelUp}
\alias{playCamelUp}
\title{Play the game CamelUp}
\usage{
playCamelUp()
}
\value{
an object representing the CamelUp app as generated by shiny::shinyApp
}
\description{
Run CamelUp in a local web browser. Running locally allows for using the app without an internet connection and running in parallel on the local computer
}
