% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{player}
\alias{player}
\title{Player object to represent each player using bets and a purse}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
player
}
\description{
Player object to represent each player using bets and a purse
}
\examples{

s <- system$new(nPlayers = 2, players = c("Michael", "Alex"))
x <- player$new("Michael", s$board)
}
\keyword{datasets}
