% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Define server logic required to draw a histogram}
\usage{
server(input, output)
}
\arguments{
\item{input}{input for shiny web app}

\item{output}{output for shiny web app}
}
\description{
Define server logic required to draw a histogram
}
