% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMove.R
\name{simulateMoveNTimes}
\alias{simulateMoveNTimes}
\title{Simulate moving N times}
\usage{
simulateMoveNTimes(g, N)
}
\arguments{
\item{g}{game object}

\item{N}{number of sims}
}
\description{
Simulate moving N times
}
