% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{space}
\alias{space}
\title{Implements spaces on the board}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
space
}
\description{
Implements spaces on the board
}
\examples{
x <- space$new()
}
\keyword{datasets}
