% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{class}
\name{stack}
\alias{stack}
\title{Implements a classic stack with push, pop and a few other methods}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stack
}
\description{
Implements a classic stack with push, pop and a few other methods
}
\examples{
x <- stack$new()
x$push(5)
y <- x$pop()
}
\keyword{datasets}
