% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBoard.R
\name{visualizeBoard}
\alias{visualizeBoard}
\title{visualizeBoard}
\usage{
visualizeBoard(g)
}
\arguments{
\item{g}{a game object}
}
\value{

}
\description{
visualizeBoard
}
\examples{
g <- Game$new(19, 3, TRUE)
visualizeBoard(g)
}
