% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{overall.bet}
\alias{overall.bet}
\title{A bet object that is placed overall on a given camel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
overall.bet
}
\description{
A bet object that is placed overall on a given camel
}
\keyword{datasets}
