% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{server}
\alias{server}
\title{Play the game CamelUp}
\usage{
server(input, output)
}
\arguments{
\item{input}{server input}

\item{output}{server output}
}
\value{
an object representing the CamelUp app as generated by shiny::shinyApp
}
\description{
Run CamelUp in a local web browser. Running locally allows for using the app without an internet connection and running in parallel on the local computer
}
