% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Space}
\alias{Space}
\title{Encapsulates a double}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor}

\item{\code{mult}}{Multiply by another Double object \itemize{
\item Paramter: other - The other Double object
\item Returns: product of the values
}}
}}

