% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Implementation.R
\docType{data}
\name{system}
\alias{system}
\title{System class that manages overall gameplay}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
system
}
\description{
System class that manages overall gameplay
}
\keyword{datasets}
